Refactored common framework dependencies to libs package
