+++ pom.xml
@@ -103,29 +103,6 @@
 
 		<!-- END -->
 
-		<!-- ONS Libraries -->
-<!--		<dependency>-->
-<!--			<groupId>uk.gov.ons.ctp.product</groupId>-->
-<!--			<artifactId>partysvc-api</artifactId>-->
-<!--			<version>10.50.7</version>-->
-<!--		</dependency>-->
-<!--		<dependency>-->
-<!--			<groupId>uk.gov.ons.ctp.product</groupId>-->
-<!--			<artifactId>samplesvc-api</artifactId>-->
-<!--			<version>10.49.19</version>-->
-<!--		</dependency>-->
-<!--		<dependency>-->
-<!--			<groupId>uk.gov.ons.ctp.product</groupId>-->
-<!--			<artifactId>reportmodule</artifactId>-->
-<!--			<version>10.49.0</version>-->
-<!--		</dependency>-->
-<!--		<dependency>-->
-<!--			<groupId>uk.gov.ons.ctp.common</groupId>-->
-<!--			<artifactId>framework</artifactId>-->
-<!--			<version>10.49.21</version>-->
-<!--		</dependency>-->
-		<!-- END -->
-
 		<!-- Third Party -->
 		<dependency>
 			<groupId>org.projectlombok</groupId>
