+++ src/main/java/uk/gov/ons/ctp/response/libs/common/rest/RestUtility.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.common.rest;
+package uk.gov.ons.ctp.response.libs.common.rest;
 
 import java.nio.charset.Charset;
 import org.springframework.http.HttpEntity;
