+++ src/main/java/uk/gov/ons/ctp/response/libs/common/state/BasicStateTransitionManager.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.common.state;
+package uk.gov.ons.ctp.response.libs.common.state;
 
 import com.godaddy.logging.Logger;
 import com.godaddy.logging.LoggerFactory;
@@ -7,7 +7,7 @@ import java.util.HashMap;
 import java.util.Map;
 import lombok.Data;
 import lombok.Getter;
-import uk.gov.ons.ctp.common.error.CTPException;
+import uk.gov.ons.ctp.response.libs.common.error.CTPException;
 
 /**
  * A Simple impl of StateTransitionManager
