+++ src/main/java/uk/gov/ons/ctp/response/sample/SampleSvcApplication.java
@@ -23,11 +23,11 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.integration.annotation.IntegrationComponentScan;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.web.client.RestTemplate;
-import uk.gov.ons.ctp.common.error.RestExceptionHandler;
-import uk.gov.ons.ctp.common.jackson.CustomObjectMapper;
-import uk.gov.ons.ctp.common.rest.RestUtility;
-import uk.gov.ons.ctp.common.state.StateTransitionManager;
-import uk.gov.ons.ctp.common.state.StateTransitionManagerFactory;
+import uk.gov.ons.ctp.response.libs.common.error.RestExceptionHandler;
+import uk.gov.ons.ctp.response.libs.common.jackson.CustomObjectMapper;
+import uk.gov.ons.ctp.response.libs.common.rest.RestUtility;
+import uk.gov.ons.ctp.response.libs.common.state.StateTransitionManager;
+import uk.gov.ons.ctp.response.libs.common.state.StateTransitionManagerFactory;
 import uk.gov.ons.ctp.response.sample.config.AppConfig;
 import uk.gov.ons.ctp.response.sample.representation.SampleSummaryDTO.SampleEvent;
 import uk.gov.ons.ctp.response.sample.representation.SampleSummaryDTO.SampleState;
