+++ src/main/java/uk/gov/ons/ctp/response/sample/config/PartySvc.java
@@ -2,7 +2,7 @@ package uk.gov.ons.ctp.response.sample.config;
 
 import lombok.Data;
 import net.sourceforge.cobertura.CoverageIgnore;
-import uk.gov.ons.ctp.common.rest.RestUtilityConfig;
+import uk.gov.ons.ctp.response.libs.common.rest.RestUtilityConfig;
 
 /** Application config bean for the connection details to the party Service. */
 @CoverageIgnore
