+++ src/main/java/uk/gov/ons/ctp/response/sample/endpoint/SampleEndpoint.java
@@ -25,9 +25,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-import uk.gov.ons.ctp.common.error.CTPException;
-import uk.gov.ons.ctp.common.error.InvalidRequestException;
-import uk.gov.ons.ctp.common.time.DateTimeUtil;
+import uk.gov.ons.ctp.response.libs.common.error.CTPException;
+import uk.gov.ons.ctp.response.libs.common.error.InvalidRequestException;
+import uk.gov.ons.ctp.response.libs.common.time.DateTimeUtil;
 import uk.gov.ons.ctp.response.libs.sample.validation.BusinessSampleUnit;
 import uk.gov.ons.ctp.response.sample.domain.model.CollectionExerciseJob;
 import uk.gov.ons.ctp.response.sample.domain.model.SampleAttributes;
