+++ src/main/java/uk/gov/ons/ctp/response/sample/service/CollectionExerciseJobService.java
@@ -5,8 +5,8 @@ import com.godaddy.logging.LoggerFactory;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.ons.ctp.common.error.CTPException;
-import uk.gov.ons.ctp.common.error.CTPException.Fault;
+import uk.gov.ons.ctp.response.libs.common.error.CTPException;
+import uk.gov.ons.ctp.response.libs.common.error.CTPException.Fault;
 import uk.gov.ons.ctp.response.sample.domain.model.CollectionExerciseJob;
 import uk.gov.ons.ctp.response.sample.domain.repository.CollectionExerciseJobRepository;
