+++ src/main/java/uk/gov/ons/ctp/response/sample/service/PartySvcClientService.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponents;
-import uk.gov.ons.ctp.common.rest.RestUtility;
+import uk.gov.ons.ctp.response.libs.common.rest.RestUtility;
 import uk.gov.ons.ctp.response.libs.party.definition.PartyCreationRequestDTO;
 import uk.gov.ons.ctp.response.libs.party.representation.PartyDTO;
 import uk.gov.ons.ctp.response.sample.config.AppConfig;
