+++ src/main/java/uk/gov/ons/ctp/response/sample/service/state/SampleSvcStateTransitionManagerFactory.java
@@ -3,9 +3,9 @@ package uk.gov.ons.ctp.response.sample.service.state;
 import java.util.HashMap;
 import java.util.Map;
 import org.springframework.stereotype.Component;
-import uk.gov.ons.ctp.common.state.BasicStateTransitionManager;
-import uk.gov.ons.ctp.common.state.StateTransitionManager;
-import uk.gov.ons.ctp.common.state.StateTransitionManagerFactory;
+import uk.gov.ons.ctp.response.libs.common.state.BasicStateTransitionManager;
+import uk.gov.ons.ctp.response.libs.common.state.StateTransitionManager;
+import uk.gov.ons.ctp.response.libs.common.state.StateTransitionManagerFactory;
 import uk.gov.ons.ctp.response.sample.representation.SampleSummaryDTO.SampleEvent;
 import uk.gov.ons.ctp.response.sample.representation.SampleSummaryDTO.SampleState;
 import uk.gov.ons.ctp.response.sample.representation.SampleUnitDTO.SampleUnitEvent;
