+++ src/test/java/uk/gov/ons/ctp/response/sample/endpoint/SampleEndpointUnitTest.java
@@ -30,10 +30,10 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.validation.BindingResult;
 import uk.gov.ons.ctp.common.FixtureHelper;
-import uk.gov.ons.ctp.common.error.CTPException;
-import uk.gov.ons.ctp.common.error.InvalidRequestException;
-import uk.gov.ons.ctp.common.error.RestExceptionHandler;
-import uk.gov.ons.ctp.common.jackson.CustomObjectMapper;
+import uk.gov.ons.ctp.response.libs.common.error.CTPException;
+import uk.gov.ons.ctp.response.libs.common.error.InvalidRequestException;
+import uk.gov.ons.ctp.response.libs.common.error.RestExceptionHandler;
+import uk.gov.ons.ctp.response.libs.common.jackson.CustomObjectMapper;
 import uk.gov.ons.ctp.response.sample.domain.model.CollectionExerciseJob;
 import uk.gov.ons.ctp.response.sample.domain.model.SampleAttributes;
 import uk.gov.ons.ctp.response.sample.domain.model.SampleSummary;
