+++ src/test/java/uk/gov/ons/ctp/response/sample/scheduled/distribution/SampleUnitDistributorTest.java
@@ -23,8 +23,8 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
-import uk.gov.ons.ctp.common.error.CTPException;
-import uk.gov.ons.ctp.common.error.CTPException.Fault;
+import uk.gov.ons.ctp.response.libs.common.error.CTPException;
+import uk.gov.ons.ctp.response.libs.common.error.CTPException.Fault;
 import uk.gov.ons.ctp.response.sample.config.AppConfig;
 import uk.gov.ons.ctp.response.sample.config.DataGrid;
 import uk.gov.ons.ctp.response.sample.domain.model.CollectionExerciseJob;
